Factor out some data constants
Specify dtypes of events and image data in a single place in
tristan.data.  Also specify a key 'pixel_index' for identifying data
pixel location data that have been converted from the Tristan event_id
form to the index of the pixel in the flattened image array.

+++ src/tristan/command_line/images/gated_images_cli.py
@@ -105,7 +105,7 @@ def main(args):
                 )
 
     num_images = open_times.size
-    bins = np.linspace(0, num_images, num_images + 1, dtype=np.uint64)
+    bins = np.linspace(0, num_images, num_images + 1, dtype=event_time_dtype)
 
     if input_nexus.exists():
         try:
